+++ TLM/TLM/TLM.csproj
@@ -135,7 +135,6 @@
     </Compile>
     <Compile Include="Constants.cs" />
     <Compile Include="Custom\PathFinding\PathfinderUpdates.cs" />
-    <Compile Include="Enum\RecklessDrivers.cs" />
     <Compile Include="Geometry\Impl\SegmentEndId.cs" />
     <Compile Include="Lifecycle\Patcher.cs" />
     <Compile Include="Lifecycle\TMPELifecycle.cs" />
