+++ TLM/TMPE.API/TMPE.API.csproj
@@ -159,6 +159,7 @@
     <Compile Include="Traffic\Enums\ParkingError.cs" />
     <Compile Include="Traffic\Enums\PriorityType.cs" />
     <Compile Include="Traffic\Enums\EmergencyBehavior.cs" />
+    <Compile Include="Traffic\Enums\RecklessDrivers.cs" />
     <Compile Include="Traffic\Enums\SetLaneArrow_Result.cs" />
     <Compile Include="Traffic\Enums\SetPrioritySignError.cs" />
     <Compile Include="Traffic\Enums\SimulationAccuracy.cs" />
